requirements.txt
google-genai
sentence-transformers
faiss-cpu
python-dotenv
streamlit
